requirements.txt
bidict==0.22.1; python_version >= '3.7'
blinker==1.7.0; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
flask==3.0.0
flask-socketio==5.3.6
h11==0.14.0; python_version >= '3.7'
inputs==0.5
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
numpy==1.26.2; python_version >= '3.10'
opencv-python==4.8.1.78
paho-mqtt==1.6.1
python-engineio==4.8.0; python_version >= '3.6'
python-socketio==5.10.0; python_version >= '3.6'
simple-websocket==1.0.0; python_version >= '3.6'
werkzeug==3.0.1; python_version >= '3.8'
wsproto==1.2.0; python_full_version >= '3.7.0'
